lambda/requirements.txt
selenium
psycopg2-binary
beautifulsoup4
webdriver-manager
chromedrivermanager